refactor: change folder name for consistency
