+++ src/commands/hashCommands/canculateHash.js
